Relocate the method
