+++ src/discord-api/utils/api-client.util.ts
@@ -1,5 +1,7 @@
-import axios, { AxiosError, isAxiosError } from 'axios'
+import axios, { Axios, AxiosError, isAxiosError } from 'axios'
 import { RouteBases, RESTJSONErrorCodes } from 'discord-api-types/v10'
+import { DiscordUserOAuth2Credentials } from 'src/discord-oauth/types'
+import { SessionUserDiscordApiClient } from '../interceptors/inject-session-user-discord-api-client/session-user-discord-api-client.class'
 
 export function createClient(accessToken: string, tokenType: string) {
   return axios.create({
@@ -8,6 +10,21 @@ export function createClient(accessToken: string, tokenType: string) {
     },
     baseURL: RouteBases.api,
   })
+}
+
+export function createSessionUserClient(
+  credentials: DiscordUserOAuth2Credentials,
+  userId: string,
+) {
+  const { accessToken, tokenType } = credentials
+  // TODO impl refresh token
+  const client = createClient(accessToken, tokenType)
+  return Object.assign<Axios, Pick<SessionUserDiscordApiClient, 'userId'>>(
+    client,
+    {
+      userId,
+    },
+  )
 }
 
 export type DiscordError = AxiosError<{
